General improvements
Standard changes, nitpicks, improved performance for polygons (they'd repeatedly get initialized, thus fuck up performance, now, instead of doing even a TextureValidID check [I've had issues with it in the past, might be a me-thing, but what I'm doing works), I'm checking if the texture_id is null, and if so, create & initialize it then set it's color [because repeatedly setting texture colors in Surface for something like this will indeed lag, best example is by testing with an image wrapper, IMO.] I also have improved variable handling, you do not need externs, [thus why I removed variables.cpp and only left inlined variables within the variables header.] and I have made the framework use constant strings ["nitpick" again, but it's still better than outright letting them reinitialize, it can be bad for performance, even if the difference relies in nanoseconds], and I have made the frame be initialized with a lambda that moves strings instead of copying them, so it's just as performant as the old method but more user friendly. I have also made various variables that were integers take only 32bit integers, just in case. I want to remind y'all that having 64 bit integers within the code will rely in 2 integers (like 32bit variables worked on 16bit machines). For the framework, something I didn't do because I didn't want the pull to be rejected for that reason, but I can do if required is use constructors for elements to avoid string copying and maybe even switch to a shared pointer method for OOP. For the engine prediction I made the cheat use the usercmd randseed instead of a MD5 one due to performance reasons and to get rid of a dependency.